requirements_entrenamiento.txt
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0 